start transport only if there is an email to send
Transports like the Swift_Transport_AbstractSmtpTransport start the connexion
during the start method.

So when running the swiftmailer:spool:send command the connexion is made to the
SMTP even if there no mail to send.

In order to avoid that the start method is called only after getting the list
of mail to send.

Swift_Transport_AbstractSmtpTransport